Add ARM Assembly code
This code is written in ARM assembly and starts with the declaration of a new section called .text. The label _start is defined as a global symbol using the .global directive.

The following instructions switch to Thumb mode by adding 1 to the program counter and branching to the resulting address. The .THUMB directive specifies that the code that follows is to be interpreted as Thumb-2 instructions.

The conn subroutine begins with a delay loop that waits for 10 microseconds using the nanosleep system call. Next, a socket is created using the socket system call with arguments that specify the address family, socket type, and protocol. The resulting file descriptor is saved in the r4 register.

The program then attempts to establish a connection to a remote host using the connect system call, passing in the socket file descriptor and a structure containing the remote host's address and port.

If the connection is successful, the program reads 9 bytes from the socket into a buffer and compares it with a password stored in memory. If the two values match, the program continues to the next step, otherwise it repeats the read and compare process.

Next, the program duplicates the socket file descriptor to standard input, output, and error using the dup2 system call, and then forks a new process using the fork system call.

In the child process, the program executes a shell using the execve system call, passing in the path to the shell executable and null pointers for the arguments and environment.

In the parent process, the program closes the socket file descriptor and returns to the beginning of the conn subroutine to attempt a new connection.

The .data section contains a definition of the sleep time used in the delay loop, the IP address and port number of the remote host, and the password that the program expects to receive from the remote host.